Restructure extension task to better support multiple version
