+++ docs/extension.md
@@ -28,22 +28,28 @@ npm install
 
 ```bash
 cd extension
-npm run build:prod
+npm run build
 ```
 
 To generate the  Azure DevOps `.vsix` extension package for testing, you'll first need to [create a publisher account](https://learn.microsoft.com/en-us/azure/devops/extend/publish/overview?view=azure-devops#create-a-publisher) on the [Visual Studio Marketplace Publishing Portal](https://marketplace.visualstudio.com/manage/createpublisher?managePageRedirect=true). After this, override your publisher ID below and generate the extension with:
 
 ```bash
-npx tfx-cli extension create --overrides-file overrides.local.json --override "{\"publisher\": \"your-publisher-id-here\"}" --json5
+npm run package -- --overrides-file overrides.local.json --override "{\"publisher\": \"your-publisher-id-here\"}"
 ```
 
 ## Installing the extension
 
 To test the extension in Azure DevOps, you'll first need to build the extension `.vsix` file (see above). After this, [publish your extension](https://learn.microsoft.com/en-us/azure/devops/extend/publish/overview?view=azure-devops#publish-your-extension), then [install your extension](https://learn.microsoft.com/en-us/azure/devops/extend/publish/overview?view=azure-devops#install-your-extension).
 
+## Running the task locally
+
+```bash
+npm start
+```
+
 ## Running the unit tests
 
 ```bash
 cd extension
-npm run test
+npm test
 ```
