+++ extension/.gitignore
@@ -1,4 +1,4 @@
 node_modules
 .taskkey
-task/**/*.js
+**/*.js
 *.vsix
