+++ extension/tasks/utils/convertPlaceholder.ts
