+++ extension/tasks/utils/extractVirtualDirectory.ts
