+++ extension/tasks/utils/resolveAzureDevOpsIdentities.ts
